% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_bands.R
\name{age_bands}
\alias{age_bands}
\title{Place age into appropriate X year band}
\usage{
age_bands(age, X = 10)
}
\arguments{
\item{age}{Age of the person.}

\item{X}{Width of year band (default: 10)}
}
\value{
The age-band that the age belongs
}
\description{
Place age into appropriate X year band
}
\examples{
age_band(3)
# "0-9"
age_band(3, X= 5)
# "0-4"

}
