% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_codons.R
\name{codon_composition}
\alias{codon_composition}
\title{Codon Composition of Sequences}
\usage{
codon_composition(orfs, orf_col, id_col)
}
\arguments{
\item{orfs}{tibble, data with cds sequences}

\item{orf_col}{character, column name containing cds}

\item{id_col}{character, column name with gene/name id}
}
\value{
tible n by x, codon counts for each sequence, n is the number
of sequences and x is the number of codons (x=61) it may contain more thant
61 in case there is a codon loke ACN
}
\description{
Codon Composition of Sequences
}
\examples{
orf <- tibble::tibble(cds= c("ACGTTT", "TTTCCC"), id=c("s0","s1"))
codon_composition(orf, orf_col = "cds", id_col = "id")
}
