% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_codons.R
\name{codon_distance}
\alias{codon_distance}
\title{Codon distance}
\usage{
codon_distance(seq1, seq2)
}
\arguments{
\item{seq2}{}
}
\value{
integer number of different codons between the sequences
}
\description{
Codon distance
}
\examples{
codon_distance("AAACCC", "ACCccc")
}
