% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_codons.R
\name{count_codons}
\alias{count_codons}
\title{Count codons in DNA sequence}
\usage{
count_codons(seq)
}
\arguments{
\item{seq}{A character.}
}
\value{
The frequency of the codons in \code{seq}
}
\description{
Count codons in DNA sequence
}
\examples{
count_codons("ACGGGG")
}
