% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{count_codons2}
\alias{count_codons2}
\title{Count codons}
\usage{
count_codons2(orf)
}
\arguments{
\item{orf}{character DNA sequence}
}
\value{
tibble with codon counts 1x64
}
\description{
Count codons
}
\examples{
count_codons("AAA")
}
