% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers.R
\name{count_kmers}
\alias{count_kmers}
\title{Count k-mers in DNA sequence}
\usage{
count_kmers(sequence, k)
}
\arguments{
\item{sequence}{str dna sequence}

\item{k}{integer k-mer length}
}
\value{
tibble with 2 columns: k-mer (string), n (number of sites)
}
\description{
Count k-mers in DNA sequence
}
\examples{
count_kmers("AAGGTTCC", k = 4)
}
