% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunning_training.R
\name{cross_fold_validation}
\alias{cross_fold_validation}
\title{Cross fold validation technique to train mRNA stability predictor}
\usage{
cross_fold_validation(k = 10)
}
\arguments{
\item{k}{integer number of folds}
}
\value{
trainControl object to use in \code{caret::train } function
}
\description{
This function generates the trainControl object to be passed to the
caret train function.
Here, I use grouped 10 fold cross validation, the motivation for this
is that the same gene (gene_id) cannot be in both groups, train and test,
at the same time. The same gene can appear multiple times in the data,
for example the human data that contains profiles for different cell types
and with differen techniques
The group variable is the `gene_id``
}
\examples{
cross_fold_validation()
}
