% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datos_preprocessed}
\alias{datos_preprocessed}
\title{Preprocessed train and test data}
\format{list
\describe{
  \item{\code{X_train}}{train predictors}
  \item{\code{X_test}}{test predictors}
  \item{\code{y_train}}{train response}
  \item{\code{y_test}}{test response}
}}
\usage{
datos_preprocessed
}
\description{
The `train_set` and `test_set`` were preprocessed with the function \code{\link{preprocessing}}
}
\keyword{datasets}
