% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_report_library.R
\name{find_first_stop_codon_position}
\alias{find_first_stop_codon_position}
\title{find the position of the 1st stop codon that occurs in frame}
\usage{
find_first_stop_codon_position(seq)
}
\arguments{
\item{seq}{dna string}
}
\value{
int of position if not position found -1
}
\description{
find the position of the 1st stop codon that occurs in frame
}
