% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_codons.R
\name{get_codons}
\alias{get_codons}
\title{Get the list of codons}
\usage{
get_codons(include_stop = TRUE)
}
\arguments{
\item{include_stop}{logical include the stop codons}
}
\value{
a vector with 64 or 61 character elements
}
\description{
Get the list of codons
}
\examples{
get_codons()
get_codons(include_stop=TRUE) # dont want stop codons
}
