% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fq_from_LIMS.R
\name{get_files}
\alias{get_files}
\title{Read Files and Saves it with SampleName
This functions reads the provided fastq file(s) and saves it to
`fout`, by merging the files. It is similar to doing `cat f1.fq f2.fq >files.fq`
in the command line}
\usage{
get_files(fqfiles, fout, ...)
}
\arguments{
\item{fqfiles}{character vector with path to fastq file(s)}

\item{fout}{name of output file}

\item{...}{}
}
\value{
file size
}
\description{
Read Files and Saves it with SampleName
This functions reads the provided fastq file(s) and saves it to
`fout`, by merging the files. It is similar to doing `cat f1.fq f2.fq >files.fq`
in the command line
}
