% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fq_from_LIMS.R
\name{get_files_for_sample}
\alias{get_files_for_sample}
\title{Get file path of fastq data}
\usage{
get_files_for_sample(sname, dir_to_files, sample_report)
}
\arguments{
\item{sname}{character SampleName to get data from}

\item{dir_to_files}{path to dir with fq}

\item{sample_report}{tibble: sample report data}
}
\value{
character file path to fastq files
}
\description{
Get file path of fastq data
}
