% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fq_from_LIMS.R
\name{get_fq_files_from_LIMS}
\alias{get_fq_files_from_LIMS}
\title{Get NGS-data (fastq) from LIMS
I use this function to move and rename data from LIMS. LIMS data is split into
different sub-fastq files for each sample. This function takes the SampleReport.csv data,
collects all the subfiles associated with a sample name, merges into a single fastq, and
saves the result into a file called as the sample name.}
\usage{
get_fq_files_from_LIMS(sreport_path, dir_to_files, output_dir = "./",
  snames, ...)
}
\arguments{
\item{sreport_path}{character, the path to sample SampleReport.csv,
this file is usually in the same directory where the data is.}

\item{dir_to_files}{The directory to the flow cell data}

\item{output_dir}{Where to save the fastq files?}

\item{snames}{optional character vector to select specific samples}

\item{...}{}
}
\description{
Get NGS-data (fastq) from LIMS
I use this function to move and rename data from LIMS. LIMS data is split into
different sub-fastq files for each sample. This function takes the SampleReport.csv data,
collects all the subfiles associated with a sample name, merges into a single fastq, and
saves the result into a file called as the sample name.
}
\examples{
\\donotrun{
# this codes will only work in the hickory cluster
# provided the data exist
sr <- "/n/analysis/Bazzini/arb/MOLNG-2424/HGCC7BGX7/Sample_Report.csv"
dir <- "/n/analysis/Bazzini/arb/MOLNG-2424/HGCC7BGX7/"
# you should use compress=FALSE, to get uncompress file
get_fq_files_from_LIMS(sr, dir, output_dir = "~/", snames = "K562_PCR_A", compress=FALSE)
}
}
