% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{initial_preproc_codon_composition}
\alias{initial_preproc_codon_composition}
\title{Initial Preprocessing function}
\usage{
initial_preproc_codon_composition(data)
}
\arguments{
\item{data:}{data to transform (i.e. train_set, test_set)}
}
\value{
A tibble: 6 x 71
}
\description{
This function computes the codon composition and an addition variable for
the log length of the open reading frame
}
