% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_mzt_utils.R
\name{load_decay_aa_codon_composition_data}
\alias{load_decay_aa_codon_composition_data}
\title{Load the codon composition anf the decay rate data
function to process and load the codon composition and decay rates for fish}
\usage{
load_decay_aa_codon_composition_data(cc_dp, decay_dp)
}
\arguments{
\item{cc_dp}{path to 190108-mzt-rna-stability/data/19-01-17-Get-ORFS-UTRS-codon-composition/sequence-data/zfish_codon_composition.csv}

\item{decay_dp}{path to 190108-mzt-rna-stability/results/19-01-11-GetDecayRateFromTimeCourse/results_data/estimated_decay_rates.csv}
}
\description{
Load the codon composition anf the decay rate data
function to process and load the codon composition and decay rates for fish
}
