% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_mzt_mdl_RNA_seq_dynamics.R
\name{make_mdl_frame}
\alias{make_mdl_frame}
\title{Make Model Frame}
\usage{
make_mdl_frame(log2FCtimecourse_dp, pls_optimality_dp, utrs_dp,
  kmers_to_keep, .sample_condition = "wt_ribo", .maternal = TRUE,
  minimum_time_point = -1)
}
\arguments{
\item{log2FCtimecourse_dp}{character to file path: data/19-02-05-FoldChangeData/data/log2FC_earlyVSlate_tidytimecourse.csv}

\item{pls_optimality_dp}{character to file path: results/19-01-18-ObtainEstimateOfCodonOptimalityPLS/results_data/pls_components_fish_genes.csv}

\item{utrs_dp}{character to file path: data/19-01-17-Get-ORFS-UTRS-codon-composition/sequence-data/fish_seq_data_cds_3utr.csv}

\item{kmers_to_keep}{character vector of k-mers to add, this predictor
represents the number of times that a k-mer is present in the 3' UTR seqs}

\item{.sample_condition}{character any of aamanit2in_polya wt_polya wt_ribo}

\item{.maternal}{logical, True only matenal genes, assume there is column is_maternal}

\item{minimum_time_point}{numeric lowest time point used for modeling}
}
\value{
tibble model frame
}
\description{
Make Model Frame
}
