% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_colours.R
\docType{data}
\name{nhsd_chart_colours}
\alias{nhsd_chart_colours}
\title{NHS Digital chart colours}
\format{An object of class \code{character} of length 7.}
\usage{
nhsd_chart_colours
}
\description{
A list of chart colours constructed from the NHS colour palette.
Contains the values "light blue", "blue", "dark blue", "warm yellow", "light grey", "grey" and "dark grey".
}
\keyword{datasets}
