% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_colours.R
\docType{data}
\name{nhsd_core_colours}
\alias{nhsd_core_colours}
\title{NHS Digital core colours}
\format{An object of class \code{character} of length 24.}
\usage{
nhsd_core_colours
}
\description{
A large selection of NHS digital core colours, selected from
https://digital.nhs.uk/about-nhs-digital/corporate-information-and-documents/nhs-digital-style-guidelines/how-we-look/colour-palette
}
\keyword{datasets}
