% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{optimality_code_embo2016}
\alias{optimality_code_embo2016}
\title{Codon Optimality Code}
\format{A data frame with 64 rows and 2 variables:
\describe{
  \item{codon}{3 nucleotide letter codon}
  \item{optimality}{optimality definition (stop, neutral, optimal, non-optimal)}
}}
\usage{
optimality_code_embo2016
}
\description{
A data set containing the codon optimality code as defined in
\href{http://emboj.embopress.org/content/35/19/2087.long}{Bazzini2016}
}
\keyword{datasets}
