% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_codons.R
\name{optimality_counts}
\alias{optimality_counts}
\title{Optimality Counts bazzini embo 2016}
\usage{
optimality_counts(seq)
}
\arguments{
\item{seq}{character, coding dna sequence}
}
\value{
tibble
}
\description{
Counts the number and percentage of optimal, non-optimal and neutral
codons accoring to bazzini et al 2016
}
\examples{
optimality_counts("AAACCCTAT")
}
