% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{prepare_train_and_test_sets}
\alias{prepare_train_and_test_sets}
\title{Preprate train and test set for modeling}
\usage{
prepare_train_and_test_sets()
}
\value{
list with train and test data
}
\description{
Preprate train and test set for modeling
}
