% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessed_data}
\alias{preprocessed_data}
\title{Preprocess data}
\usage{
preprocessed_data()
}
\value{
list with the following objects
\enumerate{
  \item \code{X_train} preprocesed training predictors
  \item \code{X_test} preprocesed testing predictors
  \item \code{y_train} training response
  \item \code{y_test} testing response
  \item \code{preprocessing_pipeline} recipe processing pipeline

}
}
\description{
This function applies the \code{\link{preprocessing}} pipeline to the training
data and preprocess train and test
}
