% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{preprocessing_pipeline}
\alias{preprocessing_pipeline}
\title{Preprocessing pipeline}
\format{recipe}
\usage{
preprocessing_pipeline
}
\description{
The trained data preprocessing recipe
}
\keyword{datasets}
