% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_colours.R
\name{scale_colour_nhsd}
\alias{scale_colour_nhsd}
\alias{scale_color_nhsd}
\title{NHSD colour scale}
\usage{
scale_colour_nhsd(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{discrete}{Whether this scale is discrete or continuous (default: TRUE)}

\item{reverse}{Whether to reverse the scale (default: FALSE)}

\item{pallette}{Which colour palette to use, can be "main", "blue", "grey" (default: "main")}
}
\value{
A colour pallette, to be added to a ggplot plot object
}
\description{
A ggplot compatible colour palette for producing NHS digital graphs.
}
\examples{
library(gapminder)

countries <- c("United Kingdom", "China", "Gambia")

gapminder \%>\%
  filter(country \%in\% countries) \%>\%
  ggplot(aes(year, lifeExp, colour=country)) +
  geom_smooth(se = FALSE) +
  labs(
    title = "Life expectancy",
    x = "Year",
    y = "Expectancy",
    caption = "Source: gapminder (2020)"
    ) +
  theme_minimal() +
  scale_colour_nhsd("blue") +
  ggsave("test.png",
         type = "cairo",
         width = 7,
         height = 4,
  )

}
