% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_colours.R
\name{scale_fill_nhsd}
\alias{scale_fill_nhsd}
\title{NHSD fill scale}
\usage{
scale_fill_nhsd(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Which colour palette to use, can be "main", "blue", "grey" (default: "main")}

\item{discrete}{Whether this scale is discrete or continuous (default: FALSE)}

\item{reverse}{Whether to reverse the scale (default: FALSE)}
}
\value{
A colour pallette
}
\description{
A ggplot compatible colour palette for producing NHS digital graphs.
}
\examples{
library(ggplot2)

ggplot(faithfuld, aes(waiting, eruptions)) +
    geom_raster(aes(fill = density)) +
    theme_minimal() +
    scale_fill_nhsd(discrete=FALSE, reverse=TRUE)

}
