% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_set}
\alias{test_set}
\title{Test data}
\format{A data frame with 7576 rows and 7 variables:
\describe{
  \item{\code{gene_id}}{character ensembl gene id}
  \item{\code{specie}}{character the specie}
  \item{\code{cell_type}}{character cell type where mRNA stability comes}
  \item{\code{datatype}}{character technique used to generate mRNA stability profile}
  \item{\code{decay_rate}}{double mRNA decay rate, scaled for each stability profile}
  \item{\code{utrlenlog}}{double log length of 3' UTR}
  \item{\code{coding}}{character ORF star to stop codon}
}}
\source{
\url{https://github.com/santiago1234/MZT-rna-stability/blob/master/results/19-04-30-PredictiveModelDecayAllSpecies/19-04-30-EDA/EDAanalysos.md}
}
\usage{
test_set
}
\description{
Data used to evaluate mRNA stability predictor, this data set
was splitted such that the same gene is not in the training and testing
sets at the same time
}
\keyword{datasets}
