% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunning_training.R
\name{train_final_model}
\alias{train_final_model}
\title{Train Final Model}
\usage{
train_final_model(X_train = datos_preprocessed$X_train,
  y_train = datos_preprocessed$y_train)
}
\arguments{
\item{X_train}{train predictors}

\item{y_train}{test response}
}
\value{
trained gradient boosted model with gaussian loss function.
}
\description{
This function trains the final model with
the final hyperparameters found after model
tunning
}
\note{
when using the output model for prediction, the
data should be a data.frame object and not a tibble, if not
the function will throw an error
}
