% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunning_training.R
\name{tune_models}
\alias{tune_models}
\title{Train Machine Learning model to predict mRNA stability}
\usage{
tune_models(ncores = 32)
}
\arguments{
\item{ncores}{}
}
\description{
Train Machine Learning model to predict mRNA stability
}
