% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_colours.R
\name{scale_colour_nhsd}
\alias{scale_colour_nhsd}
\alias{scale_color_nhsd}
\title{NHSD colour scale}
\usage{
scale_colour_nhsd(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Which colour palette to use, can be "main", "blue", "grey" (default: "main")}

\item{discrete}{Whether this scale is discrete or continuous (default: TRUE)}

\item{reverse}{Whether to reverse the scale (default: FALSE)}
}
\value{
A colour pallette, to be added to a ggplot plot object
}
\description{
A ggplot compatible colour palette for producing NHS digital graphs.
}
\examples{
library(ggplot2)

ggplot(iris, aes(Sepal.Length, Petal.Length, colour=Species)) +
  geom_point() +
  theme_minimal() +
  scale_colour_nhsd()

}
