% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunning_training.R
\name{tune_models}
\alias{tune_models}
\title{Train Machine Learning models to predict mRNA stability}
\usage{
tune_models(ncores = 32,
  output_dir_to_save_models = "results-trained-models")
}
\arguments{
\item{ncores}{integer number of cores for parallel processing. set to the
numbers of cores in your machine}

\item{output_dir_to_save_models}{character: output directory to save trained models}
}
\value{
None, all the model if not trained will be saved to \code{output_dir_to_save_models}
}
\description{
Evaluates a set of machine learning models, also perfoms hyperparameter tunning.
I follow the same approach as in chapter 10 of Applied Predictive Modeling book.
This function uses the \code{datos_preprocessed} training data for tunning,
}
\note{
This function is computational expensive and it requires parallel processing
}
